frontend/requirements.txt
streamlit
requests
Pillow # For ImageDraw and image manipulation
python-dotenv